polykybd: fw_up — fix uint8_t overflow in fw_staging_write_chunk (chunk-0 hang)
Run-5 diagnostic (FW_UP_CHUNK_NOOP_PROBE: skip only fw_staging_write_chunk)
streamed all 4458 chunks slave_ack=0xca to COMMIT, which then failed CRC as
expected (nothing staged). That localized the long-standing slave hard-lock to
fw_staging_write_chunk.

Root cause: the page-copy loop used uint8_t for `space`/`copy`. FLASH_PAGE_SIZE
is 256, so `space = FLASH_PAGE_SIZE - s_buf_fill` truncates to 0 whenever the
page buffer is empty (chunk 0 and every 256 B boundary). copy becomes 0,
remaining never decreases, and the slave spins forever inside the chunk handler
-> never replies -> master sees success:0 -> "hard lock" on chunk 0.

This explains every prior observation: dies exactly on chunk 0; the lock is an
infinite loop (not a fault, so nothing on the link); the no-op handler streams
fine; erase, core1 (NMI/Vector80/PSM), payload size and the transaction id were
all red herrings. It stayed hidden because the 05-20 baseline never called
write_chunk and the erase hang later stopped execution before chunks.

Fix: widen space/copy to uint32_t. Disable FW_UP_CHUNK_NOOP_PROBE so the real
write runs. Builds clean. Flash BOTH halves: chunks should now stream AND land
in staging, so FW_UP_COMMIT should pass CRC. Details in FW_UP_BASELINE.md (run 5).

+++ keyboards/handwired/polykybd/base/fw_staging.c
@@ -321,8 +321,12 @@ bool fw_staging_write_chunk(uint32_t offset, const uint8_t *data, uint8_t len) {
     uint8_t        remaining = len;
 
     while (remaining > 0) {
-        uint8_t space = FLASH_PAGE_SIZE - s_buf_fill;
-        uint8_t copy  = (remaining < space) ? remaining : space;
+        // NB: space/copy MUST be wider than uint8_t — FLASH_PAGE_SIZE is 256, so
+        // (FLASH_PAGE_SIZE - s_buf_fill) is 256 when the page buffer is empty and
+        // truncates to 0 in a uint8_t, making copy=0 and spinning this loop
+        // forever (the chunk-0 slave hard-lock, found 2026-05-30).
+        uint32_t space = FLASH_PAGE_SIZE - s_buf_fill;
+        uint32_t copy  = (remaining < space) ? remaining : space;
 
         memcpy(s_page_buf + s_buf_fill, src, copy);
         s_buf_fill    += copy;
